resolved changes @crashillo
