+++ app/javascript/gobierto_dashboards/modules/costs_controller.js
@@ -28,7 +28,7 @@ export class CostsController {
       const TableSecondLevel = () => import("../webapp/containers/costs/TableSecondLevel.vue");
       const TableItem = () => import("../webapp/containers/costs/TableItem.vue");
 
-      Promise.all([getRemoteData(options.costsEndpoint)]).then((rawData) => {
+      Promise.resolve(getRemoteData(options.costsEndpoint)).then((rawData) => {
         this.setGlobalVariables(rawData)
 
         const router = new VueRouter({
@@ -105,37 +105,50 @@ export class CostsController {
     //Array with all the strings that we've to convert to Number
     const amountStrings = [ 'cd_bens_i_serveis', 'cd_cost_personal', 'cost_directe_2018' , 'cost_indirecte_2018', 'cost_total_2018', 'costpers2018', 'costrestadir2018', 'cost_per_habitant', 'ingressos', 'respecte_ambit', 'taxa_o_preu_public', 'cd_serveis_exteriors', 'cd_transferencies', 'cd_equipaments', 'ingres_cost', 'subvencio']
 
-    for (let cost of rawData) {
-      for (let index = 0; index < cost.length; index++) {
-        let d = cost[index]
+    for (let index = 0; index < rawData.length; index++) {
+      let d = rawData[index]
 
-        for (let amounts = 0; amounts < amountStrings.length; amounts++) {
-          d[amountStrings[amounts]] = convertStringToNumbers(d[amountStrings[amounts]])
-        }
-
-        //Include a year column
-        if (d.cost_directe_2018) {
-          d['year'] = '2018'
-        } else if (d.cost_directe_2019) {
-          d['year'] = '2019'
-        }
-
-        d.ingressos = nanToZero(d.ingressos)
+      for (let amounts = 0; amounts < amountStrings.length; amounts++) {
+        d[amountStrings[amounts]] = convertStringToNumbers(d[amountStrings[amounts]])
       }
+
+      //Include a year column
+      if (d.cost_directe_2018) {
+        d['year'] = '2018'
+      } else if (d.cost_directe_2019) {
+        d['year'] = '2019'
+      }
+
+      d.ingressos = nanToZero(d.ingressos)
     }
 
-    const groupData = d3.nest()
-        .key(d => d.agrupacio)
-        .rollup(function(value) {
-          return {
-            count: value.length,
-            total: d3.sum(value, d => d.cost_total_2018)
-          }
-        })
-        .entries(rawData[0]);
+    let groupData = []
+    groupData = [...rawData.reduce((r, o) => {
+      const key = o.agrupacio
+
+      const item = r.get(key) || Object.assign({}, o, {
+        cost_directe_2018: 0,
+        cost_indirecte_2018: 0,
+        cost_total_2018: 0,
+        ingressos: 0,
+        respecte_ambit: 0,
+        total: 0
+      });
+
+      item.cost_directe_2018 += o.cost_directe_2018
+      item.cost_indirecte_2018 += o.cost_indirecte_2018
+      item.cost_total_2018 += o.cost_total_2018
+      item.ingressos += o.ingressos
+      //New item with the sum of values of each agrupacio
+      item.total += (o.total || 0) + 1
+      item.respecte_ambit += o.respecte_ambit
+
+      return r.set(key, item);
+    }, new Map).values()];
+    groupData = groupData.filter(element => element.agrupacio !== '')
 
     this.data = {
-      costData: rawData[0],
+      costData: rawData,
       agrupacioData: groupData
     }
   }
