+++ app/javascript/gobierto_dashboards/webapp/containers/costs/Home.vue
@@ -2,10 +2,8 @@
   <main class="gobierto-dashboards">
     <div class="pure-g gutters m_b_1">
       <Header />
-      <Distribution
-        :data="costData"
-      />
-      <Detail :data="costData" />
+      <Distribution :data="costData" />
+      <Table :data="costData" />
     </div>
   </main>
 </template>
@@ -13,13 +11,13 @@
 <script>
 import Header from './Header.vue'
 import Distribution from './Distribution.vue'
-import Detail from './Detail.vue'
+import Table from './Table.vue'
 export default {
   name: 'Home',
   components: {
     Header,
     Distribution,
-    Detail
+    Table
   },
   data() {
     return {
