+++ app/javascript/gobierto_dashboards/webapp/containers/costs/TableFirstLevel.vue
@@ -41,41 +41,10 @@ import { VueFiltersMixin } from "lib/shared"
 export default {
   name: "TableFirstLevel",
   mixins: [VueFiltersMixin],
-  props: {
-    items: {
-      type: Array,
-      default: () => []
+  data() {
+    return {
+      groupData: this.$root.$data.agrupacioData
     }
-  },
-  computed: {
-    groupData() {
-      //reduce to sum all values for agrupacio
-      let dataGroup = []
-      dataGroup = [...this.items.reduce((r, o) => {
-        const key = o.agrupacio
-
-        const item = r.get(key) || Object.assign({}, o, {
-          cost_directe_2018: 0,
-          cost_indirecte_2018: 0,
-          cost_total_2018: 0,
-          ingressos: 0,
-          respecte_ambit: 0,
-          total: 0
-        });
-
-        item.cost_directe_2018 += o.cost_directe_2018
-        item.cost_indirecte_2018 += o.cost_indirecte_2018
-        item.cost_total_2018 += o.cost_total_2018
-        item.ingressos += o.ingressos
-        //New item with the sum of values of each agrupacio
-        item.total += (o.total || 0) + 1
-        item.respecte_ambit += o.respecte_ambit
-
-        return r.set(key, item);
-      }, new Map).values()];
-      dataGroup = dataGroup.filter(element => element.agrupacio !== '')
-      return dataGroup
-    },
   }
 }
 </script>
