+++ app/javascript/gobierto_dashboards/webapp/containers/costs/TableSubHeader.vue
@@ -1,8 +1,5 @@
 <template>
-  <table
-    v-if="items.length"
-    class="gobierto-dashboards-table gobierto-dashboards-table--subheader"
-  >
+  <table class="gobierto-dashboards-table gobierto-dashboards-table--subheader">
     <tbody>
       <tr
         v-for="{ agrupacio, cost_directe_2018, cost_indirecte_2018, cost_total_2018, cost_per_habitant, ingressos, respecte_ambit } in dataGroup"
@@ -42,7 +39,7 @@ export default {
   data() {
     return {
       dataGroup: [],
-      items: this.$root.$data.costData
+      groupData: this.$root.$data.agrupacioData
     }
   },
   created() {
@@ -55,30 +52,7 @@ export default {
   },
   methods: {
     agrupacioData(id) {
-      this.dataGroup = [...this.items.reduce((r, o) => {
-        const key = o.agrupacio
-
-        const item = r.get(key) || Object.assign({}, o, {
-          cost_directe_2018: 0,
-          cost_indirecte_2018: 0,
-          cost_total_2018: 0,
-          ingressos: 0,
-          respecte_ambit: 0,
-          total: 0,
-          cost_per_habitant: 0
-        });
-
-        item.cost_directe_2018 += o.cost_directe_2018
-        item.cost_indirecte_2018 += o.cost_indirecte_2018
-        item.cost_total_2018 += o.cost_total_2018
-        item.ingressos += o.ingressos
-        item.total += (o.total || 0) + 1
-        item.respecte_ambit += o.respecte_ambit
-        item.cost_per_habitant += o.cost_per_habitant
-
-        return r.set(key, item);
-      }, new Map).values()];
-      this.dataGroup = this.dataGroup.filter(element => element.ordre_agrupacio === id)
+      this.dataGroup = this.groupData.filter(element => element.ordre_agrupacio === id)
     }
   }
 }
